Docker/requirements.txt
fastapi==0.115.6
matplotlib==3.10.0
numpy==2.2.0
opencv-python-headless
Pillow==11.0.0
Requests==2.32.3
torch==2.5.1
torchvision==0.20.1
uvicorn
httpx
python-multipart